Split path resource decoding
