+++ src/psd_tools/decoder/tagged_blocks.py
@@ -4,14 +4,14 @@ import warnings
 import collections
 import io
 
-from psd_tools.constants import (TaggedBlock, SectionDivider, PathResource,
-                                 ColorMode)
+from psd_tools.constants import TaggedBlock, SectionDivider, ColorMode
 from psd_tools.decoder.actions import (decode_descriptor, UnknownOSType,
                                        RawData)
 from psd_tools.utils import (read_fmt, unpack, read_unicode_string,
                              read_pascal_string)
 from psd_tools.decoder import decoders, layer_effects
 from psd_tools.decoder.color import decode_color
+from psd_tools.decoder.path import decode_path_resource
 from psd_tools.reader.layers import Block
 from psd_tools.debug import pretty_namedtuple
 from psd_tools.decoder import engine_data
@@ -671,43 +671,7 @@ def _decode_vector_mask_setting1(data, **kwargs):
                       "unsupported version %s" % (ver))
         return data
 
-    # Path points are 8 bits + 24 bits fixed points. Convert to float here.
-    def _decode_fixed_point(fixed_point):
-        return tuple(x / 0x01000000 for x in fixed_point)
-
-    path = []
-    path_rec = (len(data) - 8) // 26
-    while path_rec > 0:
-        selector, = read_fmt("H", fp)
-        record = {"selector": selector}
-        if selector in (PathResource.CLOSED_SUBPATH_LENGTH_RECORD,
-                PathResource.OPEN_SUBPATH_LENGTH_RECORD):
-            record["num_knot_records"], = read_fmt("H", fp)
-            fp.seek(22, io.SEEK_CUR)
-        elif selector in (
-                PathResource.CLOSED_SUBPATH_BEZIER_KNOT_LINKED,
-                PathResource.CLOSED_SUBPATH_BEZIER_KNOT_UNLINKED,
-                PathResource.OPEN_SUBPATH_BEZIER_KNOT_LINKED,
-                PathResource.OPEN_SUBPATH_BEZIER_KNOT_UNLINKED):
-            record["control_preceding_knot"] = _decode_fixed_point(
-                read_fmt("2i", fp))
-            record["anchor"] = _decode_fixed_point(read_fmt("2i", fp))
-            record["control_leaving_knot"] = _decode_fixed_point(
-                read_fmt("2i", fp))
-        elif selector == PathResource.PATH_FILL_RULE_RECORD:
-            fp.seek(24, io.SEEK_CUR)
-        elif selector == PathResource.CLIPBOARD_RECORD:
-            record["top"], record["left"], record["bottom"], record["right"],
-            record["resolution"] = _decode_fixed_point(read_fmt("5i", fp))
-            fp.seek(4, io.SEEK_CUR)
-        elif selector == PathResource.INITIAL_FILL_RULE_RECORD:
-            record["initial_fill_rule"], = read_fmt("H", fp)
-            fp.seek(22, io.SEEK_CUR)
-        else:
-            warnings.warn("Unknown path record found %s" % (selector))
-        path.append(record)
-        path_rec -= 1
-
+    path = decode_path_resource(fp.read())
     return VectorMaskSetting(
         ver, (0x01 & flags) > 0, (0x02 & flags) > 0, (0x04 & flags) > 0, path)
